M2-5553: RespondentDataSummary.utils moved to the directory utils
