+++ src/modules/Dashboard/features/RespondentData/RespondentDataSummary/Report/ReportFilters/ReportFilters.tsx
@@ -12,7 +12,7 @@ import { AutocompleteOption } from 'shared/components/FormComponents';
 
 import { FetchAnswers } from '../../RespondentDataSummary.types';
 import { useRespondentAnswers } from '../../hooks/useRespondentAnswers';
-import { getUniqueIdentifierOptions } from '../../RespondentDataSummary.utils';
+import { getUniqueIdentifierOptions } from './ReportFilters.utils';
 import { StyledFiltersContainer, StyledMoreFilters, StyledTimeText } from './ReportFilters.styles';
 import {
   FiltersChangeType,
