+++ src/modules/Dashboard/features/RespondentData/RespondentDataSummary/hooks/index.ts
@@ -1,4 +1,2 @@
 export * from './useDatavizFilters';
-export * from './useDatavizSummaryRequests';
 export * from './useDecryptedIdentifiers';
-export * from './useRespondentAnswers';
