+++ src/modules/Dashboard/features/RespondentData/RespondentDataSummary/hooks/useRespondentAnswers/index.ts
@@ -1,0 +1,1 @@
+export * from './useRespondentAnswers';
