+++ src/modules/Dashboard/features/RespondentData/RespondentDataSummary/utils/getSortedOptions.ts
@@ -1,0 +1,4 @@
+import { ItemOption } from '../../RespondentData.types';
+
+export const getSortedOptions = (options: ItemOption[]) =>
+  options.sort((a, b) => b.value - a.value);
