+++ src/modules/Dashboard/features/RespondentData/RespondentDataSummary/utils/isValueDefined.ts
@@ -1,0 +1,3 @@
+export const isValueDefined = <T extends string | number | (string | number)[] | null>(
+  value?: T,
+): value is NonNullable<T> => value !== null && value !== undefined;
